+++ src/ImageBase.h
@@ -12,7 +12,7 @@ public:
     virtual void setWidth(int w) = 0;
     virtual void setHeight(int h) = 0;
     virtual void setMaxValue(int max) = 0;
-    [[nodiscard]] virtual float getPixel(int x, int y, int c) const = 0;
+    [[nodiscard]] virtual int getPixel(int x, int y, int c) const = 0;
     virtual void setPixel(int x, int y, int c, float value) = 0;
     // virtual const std::vector<std::vector<unsigned char>> & getData() const = 0;
     virtual void setData(const std::vector<std::vector<unsigned char>>& newData) = 0;
